02-card-memory-game/requirements.txt
python-fasthtml
uvicorn>=0.29
python-multipart
sqlite-utils
huggingface-hub>=0.20.0
fasthtml-hf